requirements.txt
numpy
torch==1.13.1
matplotlib
scikit-learn
pandas
TensorFlow
huggingface_hub
python-dotenv
tiktoken
sentence-transformers
transformers
accelerate
langchain
dhg
scikit-learn
torchmetrics
torchvision
torchaudio
focal_loss_torch
openai
torch-geometric
torch_sparse
torch_scatter
tqdm